+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 - Defined output port
 - Update execution report
 - Reduce memory usage when parsing graphs from CMEM by creating temporary file 
+- Validate parameters at initialisation
 
 ## [5.1.0] 2024-12-03